support for captcha, TOTP code and TOTP gen key when detecting a captcha because the login attemp was on a new IP, a new chromium window will appear and the user will have to manually connect
when a TOTP field is detected, the user can provide the 6 number code to
the script in the shell.

you can put the TOTP generation key in the .env file and the script will
automatically generate the TOTP and use it. so you can keep the TOTP and
still automate this script.

some code improvement, more login error handling